# api/swagger.json
{
   "openapi": "3.0.0",
   "info": {
      "title": "Pilot Wellness Score API",
      "version": "2.2.0",
      "description": "\n<h2>Change History:</h2>\n<ul>\n<li><b>2.2.0</b> - 2020-08-29: Updated `RuleConfigurationValidationResponse` to include date field </li>\n<li><b>2.1.0</b> - 2020-08-27: Corrected description of `identify_high_risk_symptom_dates` in `FactResolutionExtension`; added rule configuration operation for testing </li>\n<li><b>2.0.1</b> - 2020-08-24: Corrected typo in scope required to calculate wellness </li>\n<li><b>2.0.0</b> - 2020-08-20: Updated `RuleConfiguration` and `Fact`; update `score` in `WellnessCalculationResponse` </li>\n<li><b>1.0.0</b> - 2020-08-19: Updated security schemes; removed tenant id from header </li>\n<li><b>0.1.1</b> - 2020-07-30: Updated `Error` and `ErrorResponse`; corrected string formats</li>\n<li><b>0.1.0</b> - 2020-07-29: Initial draft</li>\n</ul>\n"
   },
...
   "paths": {
...
      "/input_schemas": {
         "get": {
            "operationId": "getSchemas",
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "tags": [
               "input schemas"
            ],
            "description": "...",
            "responses": {
               "200": {
                  "description": "...",
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  },
                  "content": {
                     "application/json": {
                        "schema": {
                           "properties": {
                              "input_schemas": {
                                 "type": "array",
                                 "items": {
                                    "$ref": "#/components/schemas/InputSchema"
                                 }
                              }
                           }
                        }
                     }
                  }
               },
               "default": {
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  },
                  "description": "...",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ErrorResponse"
                        }
                     }
                  }
               }
            }
         },
         "post": {
            "operationId": "addSchema",
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "tags": [
               "input schemas"
            ],
            "description": "...",
            "requestBody": {
               "content": {
                  "application/json": {
                     "schema": {
                        "$ref": "#/components/schemas/InputSchema"
                     }
                  }
               }
            },
            "responses": {
               "201": {
                  "description": "...",
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  },
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/InputSchema"
                        }
                     }
                  }
               },
               "default": {
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  },
                  "description": "...",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ErrorResponse"
                        }
                     }
                  }
               }
            }
         }
      },
      "/input_schemas/{input_schema_id}": {
         "get": {
            "operationId": "getSchema",
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "tags": [
               "input schemas"
            ],
            "description": "...",
            "parameters": [
               {
                  "$ref": "#/components/parameters/inputSchemaIdInPath"
               }
            ],
            "responses": {
               "200": {
                  "description": "...",
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  },
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/InputSchema"
                        }
                     }
                  }
               },
               "default": {
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  },
                  "description": "...",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ErrorResponse"
                        }
                     }
                  }
               }
            }
         },
         "put": {
            "operationId": "replaceSchema",
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "tags": [
               "input schemas"
            ],
            "description": "...",
            "parameters": [
               {
                  "$ref": "#/components/parameters/inputSchemaIdInPath"
               }
            ],
            "responses": {
               "200": {
                  "description": "...",
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  },
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/InputSchema"
                        }
                     }
                  }
               },
               "default": {
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  },
                  "description": "...",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ErrorResponse"
                        }
                     }
                  }
               }
            }
         },
         "delete": {
            "operationId": "deleteSchema",
            "security": [
               {
                  "bearerAuth": []
               }
            ],
            "tags": [
               "input schemas"
            ],
            "description": "...",
            "parameters": [
               {
                  "$ref": "#/components/parameters/inputSchemaIdInPath"
               }
            ],
            "responses": {
               "200": {
                  "description": "...",
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  }
               },
               "default": {
                  "headers": {
                     "x-correlation-id": {
                        "schema": {
                           "$ref": "#/components/headers/x-correlation-id"
                        }
                     }
                  },
                  "description": "...",
                  "content": {
                     "application/json": {
                        "schema": {
                           "$ref": "#/components/schemas/ErrorResponse"
                        }
                     }
                  }
               }
            }
         }
      },
...
   },
   "components": {
      "securitySchemes": {
         "bearerAuth": {
            "type": "http",
            "scheme": "bearer",
            "bearerFormat": "JWT"
         }
      },
      "parameters": {
         "inputSchemaIdInPath": {
            "name": "input_schema_id",
            "in": "path",
            "required": true,
            "description": "...",
            "schema": {
               "type": "string"
            }
         },
...
      },
      "headers": {
...
         "x-correlation-id": {
            "schema": {
               "type": "string",
               "description": "..."
            }
         }
      },
      "schemas": {
...
         "ErrorResponse": {
            "type": "object",
            "properties": {
               "message": {
                  "type": "string"
               },
               "trace": {
                  "type": "string",
                  "description": "..."
               },
               "errors": {
                  "type": "array",
                  "description": "...",
                  "items": {
                     "$ref": "#/components/schemas/Error"
                  }
               }
            },
            "required": [
               "errors"
            ]
         },
         "Error": {
            "type": "object",
            "required": [
               "message"
            ],
            "properties": {
               "errorCode": {
                  "type": "string",
                  "description": "..."
               },
               "message": {
                  "type": "string",
                  "description": "..."
               },
               "path": {
                  "type": "string",
                  "description": "..."
               }
            }
         },
         "InputSchema": {
            "type": "object",
            "properties": {
               "id": {
                  "type": "string",
                  "description": "..."
               },
               "data": {
                  "type": "object",
                  "description": "JSON Schema - https://json-schema.org"
               }
            },
            "required": [
               "data"
            ]
         },
...
      }
   }
}